requirements.txt
docopt==0.6.2
plaid-python==1.3.0
gnupg==2.0.2
pandas==0.17.1
path.py==8.1.2
keyring==8.5.1
